gcp_project/functions/requirements.txt
google-auth==1.6.3
google-cloud-datastore==1.7.1
google-cloud-storage==1.13.0
sendgrid==6.0.5
Wand==0.5.0